Fixes AlphonsoCode/DEF#2618. Bugfix race condition for active host state.
Since, the current active namenode state is shared across webhdfs object.
So in a scenario where two almost parallel requests are made towards a same
stanby namenode, one of the requests would trigger a switchover in the
state and the other request "unaware" of the previous trigger made almost
at the same time would trigger again essentially making its request
go to the original standby namenode. The "almost same time" here means
async requests made without waiting for each others' callback and
updating the shared state as and when the callbacks are fired.

This commit fixes this bug by adding a check to distinguish between
state "seen" at the time of making the first request and the "current"
state before making a retry request. A switchover is triggered when the
two state match otherwise current state is used.